Add files via upload
The presentation slides featured here were showcased during the ICDCN'23 Conference, where I had the opportunity to share insights into our project with the academic community. The slides delve into the fundamental aspects of our research, starting from the introduction, where we outline the context and significance of our work, to the motivation behind undertaking this project.

Throughout the presentation, we provide a detail overview of the methodology employed, highlighting the techniques utilized in our study. We also delve into the challenges encountered and the strategies implemented to address them effectively.

Furthermore, the presentation showcases the results obtained from our research efforts, presenting findings, insights, and potential implications. We discuss the significance of these results in the broader context of the field and offer avenues for future research and development.